server/requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
mysqlclient==1.4.4
protobuf==3.9.2
PyMySQL==0.9.3
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.8
SQLAlchemy-serializer==1.3.2
Werkzeug==0.16.0
xlrd==1.2.0
